dagster_pipeline/requirements.txt
dagster
pandas
requests
sqlalchemy
pandera
python-dotenv
psycopg2-binary
